Fix HTTP 500 when request path points to a directory
Originally if request's path points to a directory then the server will fail
and return HTTP 500 status code. Correct behavior is to indicate issue with
request, not with the server. Thanks to this fix 404 will be returned.

Returning 5xx status code affects potential Thumbor integration to larger systems,
where 500 response may be treated as Thumbor malfunction and will cause taking it
out of load balancing pool.